requirements.txt
selenium==4.15.2
pymongo==4.6.0
flask==3.0.0
webdriver_manager==4.0.1
python-dotenv==1.0.0
requests==2.31.0
